ispr

+++ Module2/Homework2.py
@@ -11,44 +11,54 @@ while comand_lst[0]!="exit":
     if(comand_lst[0]=="pwd"):
         print(os.getcwd())
     elif(comand_lst[0]=="cd"):
-        cd_dirs = comand_lst[1].split("/" or "\ ")
-        if (cd_dirs[-1].find(".") >= 0 and cd_dirs[-1].find("..")<0):
-            print("вы ввели некоректную директорию")
+        if(len(comand_lst)>1):
+            cd_dirs = comand_lst[1].split("/" or "\ ")
+            if (cd_dirs[-1].find(".") >= 0 and cd_dirs[-1].find("..") < 0):
+                print("вы ввели некоректную директорию")
+            else:
+                copy = os.getcwd()
+                cd_path = Path(os.getcwd()) / comand_lst[1]
+                for item in cd_dirs:
+                    if Path(item).exists():
+                        cdcur_path = Path(cdcur_path) / item
+                        os.chdir(cdcur_path)
+                    else:
+                        if buf == "no":
+                            print("директории  ", comand_lst[1], "не существует")
+                            buf = input("хотите создать директорию ? (y/n)")
+                        if buf == "y":
+                            os.makedirs((comand_lst[1]))
+                            buf = "no"
+                            break
+                        elif buf == "n":
+                            os.chdir(copy)
+                            buf = "no"
+                            break
         else:
-            copy=os.getcwd()
-            cd_path = Path(os.getcwd()) / comand_lst[1]
-            for item in cd_dirs:
-                if Path(item).exists():
-                    cdcur_path=Path(cdcur_path)/item
-                    os.chdir(cdcur_path)
-                else:
-                    if buf=="no":
-                        print("директории  ", comand_lst[1], "не существует")
-                        buf = input("хотите создать директорию ? (y/n)")
-                    if buf=="y":
-                        os.makedirs((comand_lst[1]))
-                        buf="no"
-                        break
-                    elif buf=="n":
-                        os.chdir(copy)
-                        buf = "no"
-                        break
-    elif(comand_lst[0]=="touch"):
-        path_nullfile = Path(os.getcwd()) / comand_lst[1]
-        if Path(path_nullfile).exists():
-            print("Такой файл уже существует")
-        else:
-            with path_nullfile.open("w") as file:
-                print("файл создан")
-    elif(comand_lst[0]=="cat"):
-        path_file = Path(os.getcwd()) / comand_lst[1]
-        try:
-            with path_file.open("r") as file:
-                for line in file:
-                    print(line.strip())
-        except FileNotFoundError:
-            print("Файла не существует ")
+            print("вы не ввели путь")
 
+    elif(comand_lst[0]=="touch"):
+        if len(comand_lst)>1:
+            path_nullfile = Path(os.getcwd()) / comand_lst[1]
+            if Path(path_nullfile).exists():
+                print("Такой файл уже существует")
+            else:
+                with path_nullfile.open("w") as file:
+                    print("файл создан")
+        else:
+            print("вы не ввели файл")
+
+    elif(comand_lst[0]=="cat"):
+        if(len(comand_lst)>1):
+            path_file = Path(os.getcwd()) / comand_lst[1]
+            try:
+                with path_file.open("r") as file:
+                    for line in file:
+                        print(line.strip())
+            except FileNotFoundError:
+                print("Файла не существует ")
+        else:
+            print("вы не ввели файл")
     elif (comand_lst[0] == "ls"):
         directory=os.getcwd()
         filetype = os.listdir(directory)
@@ -62,11 +72,14 @@ while comand_lst[0]!="exit":
         else:
             print(filetype)
     elif (comand_lst[0] == "rm"):
-        file_del = Path(os.getcwd())/ comand_lst[1]
-        try:
-            os.remove(file_del)
-        except FileNotFoundError:
-            print("Файла не существует ")
+        if(len(comand_lst)>1):
+            file_del = Path(os.getcwd()) / comand_lst[1]
+            try:
+                os.remove(file_del)
+            except FileNotFoundError:
+                print("Файла не существует ")
+        else:
+            print("вы не ввели файл")
 
     else:
         print("Данной команды не существует ")
